Move suspense boundary for chat template footer into new component
This prevents flickering the entire chat template when fetching model params
