+++ app/src/pages/playground/PlaygroundChatTemplate.tsx
@@ -1,5 +1,9 @@
-import React, { PropsWithChildren, useCallback, useState } from "react";
-import { useLazyLoadQuery } from "react-relay";
+import React, {
+  PropsWithChildren,
+  Suspense,
+  useCallback,
+  useState,
+} from "react";
 import {
   DndContext,
   KeyboardSensor,
@@ -14,7 +18,6 @@ import {
   useSortable,
 } from "@dnd-kit/sortable";
 import { CSS } from "@dnd-kit/utilities";
-import { graphql } from "relay-runtime";
 import { css } from "@emotion/react";
 
 import {
@@ -38,31 +41,26 @@ import { usePlaygroundContext } from "@phoenix/contexts/PlaygroundContext";
 import { useChatMessageStyles } from "@phoenix/hooks/useChatMessageStyles";
 import {
   ChatMessage,
-  createOpenAIResponseFormat,
-  generateMessageId,
   PlaygroundChatTemplate as PlaygroundChatTemplateType,
   PlaygroundInstance,
 } from "@phoenix/store";
 import { assertUnreachable } from "@phoenix/typeUtils";
 import { safelyParseJSON } from "@phoenix/utils/jsonUtils";
 
-import { PlaygroundChatTemplateResponseFormatQuery } from "./__generated__/PlaygroundChatTemplateResponseFormatQuery.graphql";
 import { ChatMessageToolCallsEditor } from "./ChatMessageToolCallsEditor";
-import {
-  RESPONSE_FORMAT_PARAM_CANONICAL_NAME,
-  RESPONSE_FORMAT_PARAM_NAME,
-} from "./constants";
+import { RESPONSE_FORMAT_PARAM_CANONICAL_NAME } from "./constants";
 import {
   MessageContentRadioGroup,
   MessageMode,
 } from "./MessageContentRadioGroup";
 import { MessageRolePicker } from "./MessageRolePicker";
+import {
+  PlaygroundChatTemplateFooter,
+  PlaygroundChatTemplateFooterFallback,
+} from "./PlaygroundChatTemplateFooter";
 import { PlaygroundResponseFormat } from "./PlaygroundResponseFormat";
 import { PlaygroundTools } from "./PlaygroundTools";
-import {
-  createToolCallForProvider,
-  createToolForProvider,
-} from "./playgroundUtils";
+import { createToolCallForProvider } from "./playgroundUtils";
 import { PlaygroundInstanceProps } from "./types";
 
 const MESSAGE_Z_INDEX = 1;
@@ -83,44 +81,12 @@ export function PlaygroundChatTemplate(props: PlaygroundChatTemplateProps) {
   );
   const instances = usePlaygroundContext((state) => state.instances);
   const updateInstance = usePlaygroundContext((state) => state.updateInstance);
-  const upsertInvocationParameterInput = usePlaygroundContext(
-    (state) => state.upsertInvocationParameterInput
-  );
+
   const playgroundInstance = instances.find((instance) => instance.id === id);
   if (!playgroundInstance) {
     throw new Error(`Playground instance ${id} not found`);
   }
-  // We don't care about the model name for Azure OpenAI
-  const modelNameQueryInput =
-    playgroundInstance.model.provider !== "AZURE_OPENAI"
-      ? (playgroundInstance.model?.modelName ?? null)
-      : null;
-  const { modelInvocationParameters } =
-    useLazyLoadQuery<PlaygroundChatTemplateResponseFormatQuery>(
-      graphql`
-        query PlaygroundChatTemplateResponseFormatQuery($input: ModelsInput!) {
-          modelInvocationParameters(input: $input) {
-            __typename
-            ... on InvocationParameterBase {
-              invocationName
-              canonicalName
-            }
-          }
-        }
-      `,
-      {
-        input: {
-          providerKey: playgroundInstance.model.provider,
-          modelName: modelNameQueryInput,
-        },
-      }
-    );
 
-  const supportsResponseFormat = modelInvocationParameters?.some(
-    (p) =>
-      p.canonicalName === RESPONSE_FORMAT_PARAM_CANONICAL_NAME ||
-      p.invocationName === RESPONSE_FORMAT_PARAM_NAME
-  );
   const hasTools = playgroundInstance.tools.length > 0;
   const hasResponseFormat =
     playgroundInstance.model.invocationParameters.find(
@@ -200,89 +166,12 @@ export function PlaygroundChatTemplate(props: PlaygroundChatTemplateProps) {
         borderTopWidth="thin"
         borderBottomWidth={hasTools || hasResponseFormat ? "thin" : undefined}
       >
-        <Flex direction="row" justifyContent="end" gap="size-100">
-          {supportsResponseFormat ? (
-            <Button
-              variant="default"
-              size="compact"
-              aria-label="output schema"
-              icon={<Icon svg={<Icons.PlusOutline />} />}
-              disabled={hasResponseFormat}
-              onClick={() => {
-                upsertInvocationParameterInput({
-                  instanceId: id,
-                  invocationParameterInput: {
-                    valueJson: createOpenAIResponseFormat(),
-                    invocationName: RESPONSE_FORMAT_PARAM_NAME,
-                    canonicalName: RESPONSE_FORMAT_PARAM_CANONICAL_NAME,
-                  },
-                });
-              }}
-            >
-              Output Schema
-            </Button>
-          ) : null}
-          <Button
-            variant="default"
-            aria-label="add tool"
-            size="compact"
-            icon={<Icon svg={<Icons.PlusOutline />} />}
-            onClick={() => {
-              const patch: Partial<PlaygroundInstance> = {
-                tools: [
-                  ...playgroundInstance.tools,
-                  createToolForProvider({
-                    provider: playgroundInstance.model.provider,
-                    toolNumber: playgroundInstance.tools.length + 1,
-                  }),
-                ],
-              };
-              if (playgroundInstance.tools.length === 0) {
-                patch.toolChoice = "auto";
-              }
-              updateInstance({
-                instanceId: id,
-                patch: {
-                  tools: [
-                    ...playgroundInstance.tools,
-                    createToolForProvider({
-                      provider: playgroundInstance.model.provider,
-                      toolNumber: playgroundInstance.tools.length + 1,
-                    }),
-                  ],
-                },
-              });
-            }}
-          >
-            Tool
-          </Button>
-          <Button
-            variant="default"
-            aria-label="add message"
-            size="compact"
-            icon={<Icon svg={<Icons.PlusOutline />} />}
-            onClick={() => {
-              updateInstance({
-                instanceId: id,
-                patch: {
-                  template: {
-                    __type: "chat",
-                    messages: [
-                      ...template.messages,
-                      {
-                        id: generateMessageId(),
-                        role: "user",
-                        content: "",
-                      },
-                    ],
-                  },
-                },
-              });
-            }}
-          >
-            Message
-          </Button>
-        </Flex>
+        <Suspense fallback={<PlaygroundChatTemplateFooterFallback />}>
+          <PlaygroundChatTemplateFooter
+            instanceId={id}
+            hasResponseFormat={hasResponseFormat}
+          />
+        </Suspense>
       </View>
       {hasTools ? <PlaygroundTools {...props} /> : null}
       {hasResponseFormat ? <PlaygroundResponseFormat {...props} /> : null}
